+++ src/lib.rs
@@ -140,16 +140,17 @@ pub use tokio_tungstenite::tungstenite::Message;
 ///
 /// See the [module docs](self) for an example.
 #[derive(Debug)]
-pub struct WebSocketUpgrade {
+pub struct WebSocketUpgrade<F = DefaultOnFailedUpdgrade> {
     config: WebSocketConfig,
     /// The chosen protocol sent in the `Sec-WebSocket-Protocol` header of the response.
     protocol: Option<HeaderValue>,
     sec_websocket_key: HeaderValue,
     on_upgrade: OnUpgrade,
+    on_failed_upgrade: F,
     sec_websocket_protocol: Option<HeaderValue>,
 }
 
-impl WebSocketUpgrade {
+impl<C> WebSocketUpgrade<C> {
     /// Set the size of the internal message send queue.
     pub fn max_send_queue(mut self, max: usize) -> Self {
         self.config.max_send_queue = Some(max);
@@ -220,14 +221,23 @@ impl WebSocketUpgrade {
     where
         F: FnOnce(WebSocket) -> Fut + Send + 'static,
         Fut: Future<Output = ()> + Send + 'static,
+        C: OnFailedUpdgrade,
     {
         let on_upgrade = self.on_upgrade;
         let config = self.config;
+        let on_failed_upgrade = self.on_failed_upgrade;
 
         let protocol = self.protocol.clone();
 
         tokio::spawn(async move {
-            let upgraded = on_upgrade.await.expect("connection upgrade failed");
+            let upgraded = match on_upgrade.await {
+                Ok(upgraded) => upgraded,
+                Err(err) => {
+                    on_failed_upgrade.call(err);
+                    return;
+                }
+            };
+
             let socket =
                 WebSocketStream::from_raw_socket(upgraded, protocol::Role::Server, Some(config))
                     .await;
@@ -256,6 +266,42 @@ impl WebSocketUpgrade {
         }
 
         (StatusCode::SWITCHING_PROTOCOLS, headers).into_response()
+    }
+
+    /// Provide a callback to call if upgrading the connection fails.
+    ///
+    /// The connection upgrade is performed in a background task. If that fails this callback
+    /// will be called.
+    ///
+    /// By default any errors will be silently ignored.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use axum::response::Response;
+    /// use axum_tungstenite::WebSocketUpgrade;
+    ///
+    /// async fn handler(ws: WebSocketUpgrade) -> Response {
+    ///     ws.on_failed_upgrade(|error| {
+    ///         report_error(error);
+    ///     })
+    ///     .on_upgrade(|socket| async { /* ... */ })
+    /// }
+    /// #
+    /// # fn report_error(_: hyper::Error) {}
+    /// ```
+    pub fn on_failed_upgrade<C2>(self, callback: C2) -> WebSocketUpgrade<C2>
+    where
+        C2: OnFailedUpdgrade,
+    {
+        WebSocketUpgrade {
+            config: self.config,
+            protocol: self.protocol,
+            sec_websocket_key: self.sec_websocket_key,
+            on_upgrade: self.on_upgrade,
+            on_failed_upgrade: callback,
+            sec_websocket_protocol: self.sec_websocket_protocol,
+        }
     }
 }
 
@@ -298,6 +344,7 @@ where
             protocol: None,
             sec_websocket_key,
             on_upgrade,
+            on_failed_upgrade: DefaultOnFailedUpdgrade,
             sec_websocket_protocol,
         })
     }
@@ -395,6 +442,35 @@ fn sign(key: &[u8]) -> HeaderValue {
     sha1.update(&b"258EAFA5-E914-47DA-95CA-C5AB0DC85B11"[..]);
     let b64 = Bytes::from(base64::encode(sha1.finalize()));
     HeaderValue::from_maybe_shared(b64).expect("base64 is a valid value")
+}
+
+/// What to do when a connection upgrade fails.
+///
+/// See [`WebSocketUpgrade::on_failed_upgrade`] for more details.
+pub trait OnFailedUpdgrade: Send + 'static {
+    /// Call the callback.
+    fn call(self, error: hyper::Error);
+}
+
+impl<F> OnFailedUpdgrade for F
+where
+    F: FnOnce(hyper::Error) + Send + 'static,
+{
+    fn call(self, error: hyper::Error) {
+        self(error)
+    }
+}
+
+/// The default `OnFailedUpdgrade` used by `WebSocketUpgrade`.
+///
+/// It simply ignores the error.
+#[non_exhaustive]
+#[derive(Debug)]
+pub struct DefaultOnFailedUpdgrade;
+
+impl OnFailedUpdgrade for DefaultOnFailedUpdgrade {
+    #[inline]
+    fn call(self, _error: hyper::Error) {}
 }
 
 pub mod rejection {